Add testing for access control
