--- packages/cli/test/integration/project-role-access.api.test.ts
@@ -1,1657 +1,0 @@
-import {
-	createTeamProject,
-	getPersonalProject,
-	linkUserToProject,
-	createWorkflow,
-	randomCredentialPayload,
-	testDb,
-	mockInstance,
-} from '@n8n/backend-test-utils';
-import type { Project, User, Role } from '@n8n/db';
-import { GLOBAL_MEMBER_ROLE, GLOBAL_OWNER_ROLE } from '@n8n/db';
-import {
-	PROJECT_ADMIN_ROLE_SLUG,
-	PROJECT_EDITOR_ROLE_SLUG,
-	PROJECT_VIEWER_ROLE_SLUG,
-	PROJECT_OWNER_ROLE_SLUG,
-} from '@n8n/permissions';
-
-import { UserManagementMailer } from '@/user-management/email';
-
-import { createCustomRoleWithScopeSlugs, cleanupRolesAndScopes } from './shared/db/roles';
-import { createAdmin, createOwner, createMember } from './shared/db/users';
-import type { SuperAgentTest } from './shared/types';
-import * as utils from './shared/utils/';
-
-/**
- * Comprehensive integration tests for project role access through Web API
- * Tests both built-in roles and custom roles for workflow and credential access
- */
-
-const testServer = utils.setupTestServer({
-	endpointGroups: ['workflows', 'credentials'],
-	enabledFeatures: ['feat:sharing', 'feat:customRoles'],
-	quotas: {
-		'quota:maxTeamProjects': -1,
-	},
-});
-
-let owner: User;
-let admin: User;
-let member1: User;
-let member2: User;
-let member3: User;
-
-// Projects for different test scenarios
-let teamProjectA: Project;
-let teamProjectB: Project;
-let ownerPersonalProject: Project;
-let member1PersonalProject: Project;
-
-// Custom roles for testing (using existing scope system)
-let customWorkflowReader: Role;
-let customWorkflowWriter: Role;
-let customCredentialReader: Role;
-let customCredentialWriter: Role;
-let customWorkflowWriteOnly: Role;
-let customWorkflowDeleteOnly: Role;
-let customCredentialWriteOnly: Role;
-let customCredentialDeleteOnly: Role;
-
-// Additional specialized custom roles for CHUNK 3 testing
-let customMixedReader: Role;
-
-// Authentication agents
-let ownerAgent: SuperAgentTest;
-let adminAgent: SuperAgentTest;
-let member1Agent: SuperAgentTest;
-let member2Agent: SuperAgentTest;
-let member3Agent: SuperAgentTest;
-
-describe('Project Role Access API Integration Tests', () => {
-	beforeAll(async () => {
-		mockInstance(UserManagementMailer, {
-			invite: jest.fn(),
-			passwordReset: jest.fn(),
-		});
-
-		// Create standard users
-		owner = await createOwner();
-		admin = await createAdmin();
-		member1 = await createMember();
-		member2 = await createMember();
-		member3 = await createMember();
-
-		// Get personal projects
-		ownerPersonalProject = await getPersonalProject(owner);
-		member1PersonalProject = await getPersonalProject(member1);
-
-		// Create team projects for testing
-		teamProjectA = await createTeamProject('Team Project A', owner);
-		teamProjectB = await createTeamProject('Team Project B', owner);
-
-		// Create authentication agents
-		ownerAgent = testServer.authAgentFor(owner);
-		adminAgent = testServer.authAgentFor(admin);
-		member1Agent = testServer.authAgentFor(member1);
-		member2Agent = testServer.authAgentFor(member2);
-		member3Agent = testServer.authAgentFor(member3);
-
-		// Create custom roles using predefined scope slugs from the permissions system
-		customWorkflowReader = await createCustomRoleWithScopeSlugs(
-			['workflow:read', 'workflow:list'], // Using existing scope slugs
-			{
-				roleType: 'project',
-				displayName: 'Custom Workflow Reader',
-				description: 'Can only read workflows',
-			},
-		);
-
-		customWorkflowWriter = await createCustomRoleWithScopeSlugs(
-			['workflow:read', 'workflow:list', 'workflow:create', 'workflow:update'],
-			{
-				roleType: 'project',
-				displayName: 'Custom Workflow Writer',
-				description: 'Can read and write workflows',
-			},
-		);
-
-		customCredentialReader = await createCustomRoleWithScopeSlugs(
-			['credential:read', 'credential:list'],
-			{
-				roleType: 'project',
-				displayName: 'Custom Credential Reader',
-				description: 'Can only read credentials',
-			},
-		);
-
-		customCredentialWriter = await createCustomRoleWithScopeSlugs(
-			['credential:read', 'credential:list', 'credential:create', 'credential:update'],
-			{
-				roleType: 'project',
-				displayName: 'Custom Credential Writer',
-				description: 'Can read and write credentials',
-			},
-		);
-
-		// Create specialized roles for edge case testing
-		customWorkflowWriteOnly = await createCustomRoleWithScopeSlugs(
-			['workflow:create', 'workflow:update'], // Write permissions only, no read
-			{
-				roleType: 'project',
-				displayName: 'Custom Workflow Write-Only',
-				description: 'Can only write workflows (no read access)',
-			},
-		);
-
-		customWorkflowDeleteOnly = await createCustomRoleWithScopeSlugs(
-			['workflow:delete'], // Delete only, no read/write
-			{
-				roleType: 'project',
-				displayName: 'Custom Workflow Delete-Only',
-				description: 'Can only delete workflows (no read/write access)',
-			},
-		);
-
-		customCredentialWriteOnly = await createCustomRoleWithScopeSlugs(
-			['credential:create', 'credential:update'], // Write permissions only, no read
-			{
-				roleType: 'project',
-				displayName: 'Custom Credential Write-Only',
-				description: 'Can only write credentials (no read access)',
-			},
-		);
-
-		customCredentialDeleteOnly = await createCustomRoleWithScopeSlugs(
-			['credential:delete'], // Delete only, no read/write
-			{
-				roleType: 'project',
-				displayName: 'Custom Credential Delete-Only',
-				description: 'Can only delete credentials (no read/write access)',
-			},
-		);
-
-		// Additional specialized custom roles for comprehensive testing
-		customMixedReader = await createCustomRoleWithScopeSlugs(
-			['workflow:read', 'workflow:list', 'credential:read', 'credential:list'],
-			{
-				roleType: 'project',
-				displayName: 'Custom Mixed Reader',
-				description: 'Can read both workflows and credentials',
-			},
-		);
-	});
-
-	beforeEach(async () => {
-		await testDb.truncate(['SharedWorkflow', 'SharedCredentials']);
-	});
-
-	afterAll(async () => {
-		await cleanupRolesAndScopes();
-	});
-
-	describe('🔧 Foundation Setup Tests', () => {
-		test('should have created all required test users', () => {
-			expect(owner).toBeDefined();
-			expect(admin).toBeDefined();
-			expect(member1).toBeDefined();
-			expect(member2).toBeDefined();
-			expect(member3).toBeDefined();
-
-			expect(owner.role.slug).toBe(GLOBAL_OWNER_ROLE.slug);
-			expect(member1.role.slug).toBe(GLOBAL_MEMBER_ROLE.slug);
-		});
-
-		test('should have created all required test projects', () => {
-			expect(teamProjectA).toBeDefined();
-			expect(teamProjectB).toBeDefined();
-			expect(ownerPersonalProject).toBeDefined();
-			expect(member1PersonalProject).toBeDefined();
-
-			expect(teamProjectA.name).toBe('Team Project A');
-			expect(teamProjectB.name).toBe('Team Project B');
-		});
-
-		test('should have created all required custom roles with correct scopes', () => {
-			expect(customWorkflowReader).toBeDefined();
-			expect(customWorkflowWriter).toBeDefined();
-			expect(customCredentialReader).toBeDefined();
-			expect(customCredentialWriter).toBeDefined();
-			expect(customWorkflowWriteOnly).toBeDefined();
-			expect(customWorkflowDeleteOnly).toBeDefined();
-			expect(customCredentialWriteOnly).toBeDefined();
-			expect(customCredentialDeleteOnly).toBeDefined();
-			expect(customMixedReader).toBeDefined();
-
-			// Verify custom roles are not system roles
-			expect(customWorkflowReader.systemRole).toBe(false);
-			expect(customWorkflowWriter.systemRole).toBe(false);
-			expect(customMixedReader.systemRole).toBe(false);
-
-			// Verify scope counts match expectations
-			expect(customWorkflowReader.scopes).toHaveLength(2); // read, list
-			expect(customWorkflowWriter.scopes).toHaveLength(4); // read, list, create, update
-			expect(customWorkflowWriteOnly.scopes).toHaveLength(2); // create, update
-			expect(customWorkflowDeleteOnly.scopes).toHaveLength(1); // delete only
-			expect(customCredentialReader.scopes).toHaveLength(2); // read, list
-			expect(customCredentialWriter.scopes).toHaveLength(4); // read, list, create, update
-			expect(customMixedReader.scopes).toHaveLength(4); // workflow read/list + credential read/list
-		});
-
-		test('should verify built-in project role constants are available', () => {
-			expect(PROJECT_ADMIN_ROLE_SLUG).toBe('project:admin');
-			expect(PROJECT_EDITOR_ROLE_SLUG).toBe('project:editor');
-			expect(PROJECT_VIEWER_ROLE_SLUG).toBe('project:viewer');
-			expect(PROJECT_OWNER_ROLE_SLUG).toBe('project:personalOwner');
-		});
-
-		test('should have functional authentication agents', () => {
-			expect(ownerAgent).toBeDefined();
-			expect(adminAgent).toBeDefined();
-			expect(member1Agent).toBeDefined();
-			expect(member2Agent).toBeDefined();
-			expect(member3Agent).toBeDefined();
-		});
-	});
-
-	describe('🔐 Built-in Role Matrix Testing', () => {
-		describe('Project Admin Role - Full Project Access', () => {
-			beforeEach(async () => {
-				// Link member1 to teamProjectA as project admin
-				await linkUserToProject(member1, teamProjectA, PROJECT_ADMIN_ROLE_SLUG);
-			});
-
-			test('project admin should have full workflow access', async () => {
-				// Create a workflow in the project via HTTP endpoint (project admin has workflow:create scope)
-				const workflowPayload = {
-					name: 'Test Admin Workflow',
-					active: false,
-					nodes: [
-						{
-							id: 'uuid-1234',
-							parameters: {},
-							name: 'Start',
-							type: 'n8n-nodes-base.start',
-							typeVersion: 1,
-							position: [240, 300],
-						},
-					],
-					connections: {},
-					staticData: null,
-					settings: {
-						saveExecutionProgress: true,
-					},
-					projectId: teamProjectA.id,
-				};
-
-				const adminWorkflowResponse = await member1Agent
-					.post('/workflows')
-					.send(workflowPayload)
-					.expect(200);
-
-				const workflow = adminWorkflowResponse.body.data;
-
-				// Test workflow list access
-				const listResponse = await member1Agent.get('/workflows').expect(200);
-				expect(listResponse.body.data).toHaveLength(1);
-				expect(listResponse.body.data[0].name).toBe('Test Admin Workflow');
-
-				// Test workflow read access
-				const getResponse = await member1Agent.get(`/workflows/${workflow.id}`).expect(200);
-				expect(getResponse.body.data.name).toBe('Test Admin Workflow');
-
-				// Test workflow update access
-				const updateResponse = await member1Agent
-					.patch(`/workflows/${workflow.id}`)
-					.send({ name: 'Updated Admin Workflow', versionId: workflow.versionId })
-					.expect(200);
-				expect(updateResponse.body.data.name).toBe('Updated Admin Workflow');
-
-				// Test workflow delete access
-				await member1Agent.post(`/workflows/${workflow.id}/archive`).send().expect(200);
-				await member1Agent.delete(`/workflows/${workflow.id}`).send().expect(200);
-			});
-
-			test('project admin should have full credential access', async () => {
-				// Create a credential payload
-				const credentialPayload = randomCredentialPayload();
-
-				// Test credential create access
-				const adminCredentialResponse = await member1Agent
-					.post('/credentials')
-					.send({ ...credentialPayload, projectId: teamProjectA.id })
-					.expect(200);
-
-				const credentialId = adminCredentialResponse.body.data.id;
-
-				// Test credential list access
-				const listResponse = await member1Agent.get('/credentials').expect(200);
-				expect(listResponse.body.data).toHaveLength(1);
-				expect(listResponse.body.data[0].name).toBe(credentialPayload.name);
-
-				// Test credential read access
-				const getResponse = await member1Agent.get(`/credentials/${credentialId}`).expect(200);
-				expect(getResponse.body.data.name).toBe(credentialPayload.name);
-
-				// Test credential update access
-				const updateResponse = await member1Agent
-					.patch(`/credentials/${credentialId}`)
-					.send({ ...credentialPayload, name: 'Updated Admin Credential' })
-					.expect(200);
-				expect(updateResponse.body.data.name).toBe('Updated Admin Credential');
-
-				// Test credential delete access
-				await member1Agent.delete(`/credentials/${credentialId}`).expect(200);
-			});
-		});
-
-		describe('Project Editor Role - Full CRUD Access', () => {
-			beforeEach(async () => {
-				// Link member2 to teamProjectA as project editor
-				await linkUserToProject(member2, teamProjectA, PROJECT_EDITOR_ROLE_SLUG);
-			});
-
-			test('project editor should have full workflow CRUD access', async () => {
-				// Create a workflow in the project via HTTP endpoint (project editor has workflow:create scope)
-				const workflowPayload = {
-					name: 'Test Editor Workflow',
-					active: false,
-					nodes: [
-						{
-							id: 'uuid-1234',
-							parameters: {},
-							name: 'Start',
-							type: 'n8n-nodes-base.start',
-							typeVersion: 1,
-							position: [240, 300],
-						},
-					],
-					connections: {},
-					staticData: null,
-					settings: {
-						saveExecutionProgress: true,
-					},
-					projectId: teamProjectA.id,
-				};
-
-				const editorWorkflowResponse = await member2Agent
-					.post('/workflows')
-					.send(workflowPayload)
-					.expect(200);
-
-				const workflow = editorWorkflowResponse.body.data;
-
-				// Test workflow list access
-				const listResponse = await member2Agent.get('/workflows').expect(200);
-				expect(listResponse.body.data).toHaveLength(1);
-
-				// Test workflow read access
-				const getResponse = await member2Agent.get(`/workflows/${workflow.id}`).expect(200);
-				expect(getResponse.body.data.name).toBe('Test Editor Workflow');
-
-				// Test workflow update access
-				const updateResponse = await member2Agent
-					.patch(`/workflows/${workflow.id}`)
-					.send({ name: 'Updated Editor Workflow', versionId: workflow.versionId })
-					.expect(200);
-				expect(updateResponse.body.data.name).toBe('Updated Editor Workflow');
-
-				// Test workflow delete access (should succeed - editors have delete permission)
-				await member2Agent.post(`/workflows/${workflow.id}/archive`).send().expect(200);
-				await member2Agent.delete(`/workflows/${workflow.id}`).send().expect(200);
-			});
-
-			test('project editor should have full credential CRUD access', async () => {
-				// Create a credential payload
-				const credentialPayload = randomCredentialPayload();
-
-				// Test credential create access
-				const editorCredentialResponse = await member2Agent
-					.post('/credentials')
-					.send({ ...credentialPayload, projectId: teamProjectA.id })
-					.expect(200);
-
-				const credentialId = editorCredentialResponse.body.data.id;
-
-				// Test credential list access
-				const listResponse = await member2Agent.get('/credentials').expect(200);
-				expect(listResponse.body.data).toHaveLength(1);
-
-				// Test credential read access
-				const getResponse = await member2Agent.get(`/credentials/${credentialId}`).expect(200);
-				expect(getResponse.body.data.name).toBe(credentialPayload.name);
-
-				// Test credential update access
-				const updateResponse = await member2Agent
-					.patch(`/credentials/${credentialId}`)
-					.send({ ...credentialPayload, name: 'Updated Editor Credential' })
-					.expect(200);
-				expect(updateResponse.body.data.name).toBe('Updated Editor Credential');
-
-				// Test credential delete access (should succeed - editors have delete permission)
-				await member2Agent.delete(`/credentials/${credentialId}`).expect(200);
-			});
-		});
-
-		describe('Project Viewer Role - Read-Only Access (No Create/Update/Delete)', () => {
-			beforeEach(async () => {
-				// Link member3 to teamProjectA as project viewer
-				await linkUserToProject(member3, teamProjectA, PROJECT_VIEWER_ROLE_SLUG);
-			});
-
-			test('project viewer should have read-only workflow access (no create/update/delete permissions)', async () => {
-				// Create a workflow in the project
-				const workflow = await createWorkflow({ name: 'Test Viewer Workflow' }, teamProjectA);
-
-				// Test workflow list access
-				const listResponse = await member3Agent.get('/workflows').expect(200);
-				expect(listResponse.body.data).toHaveLength(1);
-				expect(listResponse.body.data[0].name).toBe('Test Viewer Workflow');
-
-				// Test workflow read access
-				const getResponse = await member3Agent.get(`/workflows/${workflow.id}`).expect(200);
-				expect(getResponse.body.data.name).toBe('Test Viewer Workflow');
-
-				// Test workflow create access (should be forbidden)
-				const workflowPayload = {
-					name: 'New Viewer Workflow',
-					nodes: [
-						{
-							id: 'uuid-1234',
-							parameters: {},
-							name: 'Start',
-							type: 'n8n-nodes-base.start',
-							typeVersion: 1,
-							position: [240, 300],
-						},
-					],
-					connections: {},
-					staticData: null,
-					settings: {
-						saveExecutionProgress: true,
-					},
-					projectId: teamProjectA.id,
-				};
-
-				// Project viewer doesn't have workflow:create permissions in this team project
-				await member3Agent.post('/workflows').send(workflowPayload).expect(400);
-
-				// Test workflow update access (should be forbidden)
-				await member3Agent
-					.patch(`/workflows/${workflow.id}`)
-					.send({ name: 'Updated Viewer Workflow', versionId: workflow.versionId })
-					.expect(403);
-
-				// Test workflow delete access (should be forbidden)
-				await member3Agent.post(`/workflows/${workflow.id}/archive`).send().expect(403);
-			});
-
-			test('project viewer should have read-only credential access (no create/update/delete permissions)', async () => {
-				// Create a credential via owner first
-				const credentialPayload = randomCredentialPayload();
-				const ownerCredentialResponse = await ownerAgent
-					.post('/credentials')
-					.send({ ...credentialPayload, projectId: teamProjectA.id })
-					.expect(200);
-
-				const credentialId = ownerCredentialResponse.body.data.id;
-
-				// Test credential list access
-				const listResponse = await member3Agent.get('/credentials').expect(200);
-				expect(listResponse.body.data).toHaveLength(1);
-				expect(listResponse.body.data[0].name).toBe(credentialPayload.name);
-
-				// Test credential read access
-				const getResponse = await member3Agent.get(`/credentials/${credentialId}`).expect(200);
-				expect(getResponse.body.data.name).toBe(credentialPayload.name);
-
-				// Test credential create access (should be forbidden)
-				await member3Agent
-					.post('/credentials')
-					.send({ ...randomCredentialPayload(), projectId: teamProjectA.id })
-					.expect(400);
-
-				// Test credential update access (should be forbidden)
-				await member3Agent
-					.patch(`/credentials/${credentialId}`)
-					.send({ ...credentialPayload, name: 'Updated Viewer Credential' })
-					.expect(403);
-
-				// Test credential delete access (should be forbidden)
-				await member3Agent.delete(`/credentials/${credentialId}`).expect(403);
-			});
-		});
-
-		describe('Personal Project Owner Role - Personal Project Full Access', () => {
-			test('personal project owner should have full access in their own project', async () => {
-				// Create a workflow in member1's personal project via HTTP endpoint (personal owner has workflow:create scope)
-				const workflowPayload = {
-					name: 'Test Personal Workflow',
-					active: false,
-					nodes: [
-						{
-							id: 'uuid-1234',
-							parameters: {},
-							name: 'Start',
-							type: 'n8n-nodes-base.start',
-							typeVersion: 1,
-							position: [240, 300],
-						},
-					],
-					connections: {},
-					staticData: null,
-					settings: {
-						saveExecutionProgress: true,
-					},
-					projectId: member1PersonalProject.id,
-				};
-
-				const personalWorkflowResponse = await member1Agent
-					.post('/workflows')
-					.send(workflowPayload)
-					.expect(200);
-
-				const workflow = personalWorkflowResponse.body.data;
-
-				// Test workflow list access
-				const listResponse = await member1Agent.get('/workflows').expect(200);
-				expect(listResponse.body.data).toHaveLength(1);
-				expect(listResponse.body.data[0].name).toBe('Test Personal Workflow');
-
-				// Test workflow operations
-				await member1Agent.get(`/workflows/${workflow.id}`).expect(200);
-				await member1Agent
-					.patch(`/workflows/${workflow.id}`)
-					.send({ name: 'Updated Personal Workflow', versionId: workflow.versionId })
-					.expect(200);
-				await member1Agent.post(`/workflows/${workflow.id}/archive`).send().expect(200);
-				await member1Agent.delete(`/workflows/${workflow.id}`).send().expect(200);
-
-				// Test credential operations
-				const credentialPayload = randomCredentialPayload();
-				const personalCredentialResponse = await member1Agent
-					.post('/credentials')
-					.send({ ...credentialPayload, projectId: member1PersonalProject.id })
-					.expect(200);
-
-				const credentialId = personalCredentialResponse.body.data.id;
-				await member1Agent.get(`/credentials/${credentialId}`).expect(200);
-				await member1Agent
-					.patch(`/credentials/${credentialId}`)
-					.send({ ...credentialPayload, name: 'Updated Personal Credential' })
-					.expect(200);
-				await member1Agent.delete(`/credentials/${credentialId}`).expect(200);
-			});
-		});
-	});
-
-	describe('🎯 CHUNK 3: Custom Role Functionality Tests', () => {
-		describe('Custom Role Creation & Validation Tests', () => {
-			test('should validate single-scope custom workflow roles work correctly', async () => {
-				// Link member1 with workflow-read-only role
-				await linkUserToProject(member1, teamProjectA, customWorkflowReader.slug);
-
-				// Create workflow via owner first
-				const workflow = await createWorkflow({ name: 'Single Scope Test Workflow' }, teamProjectA);
-
-				// Test allowed operations: read and list
-				const listResponse = await member1Agent.get('/workflows').expect(200);
-				expect(listResponse.body.data).toHaveLength(1);
-
-				const getResponse = await member1Agent.get(`/workflows/${workflow.id}`).expect(200);
-				expect(getResponse.body.data.name).toBe('Single Scope Test Workflow');
-
-				// Test forbidden operations: create, update, delete
-				const workflowPayload = {
-					name: 'Forbidden Workflow',
-					nodes: [
-						{
-							id: 'uuid-1234',
-							parameters: {},
-							name: 'Start',
-							type: 'n8n-nodes-base.start',
-							typeVersion: 1,
-							position: [240, 300],
-						},
-					],
-					connections: {},
-					projectId: teamProjectA.id,
-				};
-
-				// Should not be able to create
-				await member1Agent.post('/workflows').send(workflowPayload).expect(400);
-
-				// Should not be able to update
-				await member1Agent
-					.patch(`/workflows/${workflow.id}`)
-					.send({ name: 'Updated Name', versionId: workflow.versionId })
-					.expect(403);
-
-				// Should not be able to delete
-				await member1Agent.delete(`/workflows/${workflow.id}`).expect(403);
-			});
-
-			test('should validate single-scope custom credential roles work correctly', async () => {
-				// Link member2 with credential-read-only role
-				await linkUserToProject(member2, teamProjectA, customCredentialReader.slug);
-
-				// Create credential via owner first
-				const credentialPayload = randomCredentialPayload();
-				const ownerCredentialResponse = await ownerAgent
-					.post('/credentials')
-					.send({ ...credentialPayload, projectId: teamProjectA.id })
-					.expect(200);
-
-				const credentialId = ownerCredentialResponse.body.data.id;
-
-				// Test allowed operations: read and list
-				const listResponse = await member2Agent.get('/credentials').expect(200);
-				expect(listResponse.body.data).toHaveLength(1);
-
-				const getResponse = await member2Agent.get(`/credentials/${credentialId}`).expect(200);
-				expect(getResponse.body.data.name).toBe(credentialPayload.name);
-
-				// Test forbidden operations: create, update, delete
-				const newCredentialPayload = randomCredentialPayload();
-
-				// Should not be able to create
-				await member2Agent
-					.post('/credentials')
-					.send({ ...newCredentialPayload, projectId: teamProjectA.id })
-					.expect(400);
-
-				// Should not be able to update
-				await member2Agent
-					.patch(`/credentials/${credentialId}`)
-					.send({ ...credentialPayload, name: 'Updated Name' })
-					.expect(403);
-
-				// Should not be able to delete
-				await member2Agent.delete(`/credentials/${credentialId}`).expect(403);
-			});
-
-			test('should validate multi-scope combinations work correctly', async () => {
-				// Link member3 with workflow writer role (read + list + create + update)
-				await linkUserToProject(member3, teamProjectA, customWorkflowWriter.slug);
-
-				// Test create operation
-				const workflowPayload = {
-					name: 'Multi-scope Test Workflow',
-					active: false,
-					nodes: [
-						{
-							id: 'uuid-1234',
-							parameters: {},
-							name: 'Start',
-							type: 'n8n-nodes-base.start',
-							typeVersion: 1,
-							position: [240, 300],
-						},
-					],
-					connections: {},
-					projectId: teamProjectA.id,
-				};
-
-				const createResponse = await member3Agent
-					.post('/workflows')
-					.send(workflowPayload)
-					.expect(200);
-
-				const workflow = createResponse.body.data;
-
-				// Test read operations
-				const listResponse = await member3Agent.get('/workflows').expect(200);
-				expect(listResponse.body.data).toHaveLength(1);
-
-				const getResponse = await member3Agent.get(`/workflows/${workflow.id}`).expect(200);
-				expect(getResponse.body.data.name).toBe('Multi-scope Test Workflow');
-
-				// Test update operation
-				const updateResponse = await member3Agent
-					.patch(`/workflows/${workflow.id}`)
-					.send({ name: 'Updated Multi-scope Workflow', versionId: workflow.versionId })
-					.expect(200);
-
-				expect(updateResponse.body.data.name).toBe('Updated Multi-scope Workflow');
-
-				// Test forbidden operation: delete (not in scope)
-				await member3Agent.delete(`/workflows/${workflow.id}`).expect(403);
-			});
-
-			test('should validate mixed workflow/credential permissions work correctly', async () => {
-				// Link member1 with mixed reader role (workflow + credential read permissions)
-				await linkUserToProject(member1, teamProjectB, customMixedReader.slug);
-
-				// Create workflow via owner
-				const workflow = await createWorkflow({ name: 'Mixed Reader Test Workflow' }, teamProjectB);
-
-				// Create credential via owner
-				const credentialPayload = randomCredentialPayload();
-				const ownerCredentialResponse = await ownerAgent
-					.post('/credentials')
-					.send({ ...credentialPayload, projectId: teamProjectB.id })
-					.expect(200);
-
-				const credentialId = ownerCredentialResponse.body.data.id;
-
-				// Test workflow read permissions
-				const workflowListResponse = await member1Agent.get('/workflows').expect(200);
-				expect(workflowListResponse.body.data).toHaveLength(1);
-
-				const workflowGetResponse = await member1Agent.get(`/workflows/${workflow.id}`).expect(200);
-				expect(workflowGetResponse.body.data.name).toBe('Mixed Reader Test Workflow');
-
-				// Test credential read permissions
-				const credentialListResponse = await member1Agent.get('/credentials').expect(200);
-				expect(credentialListResponse.body.data).toHaveLength(1);
-
-				const credentialGetResponse = await member1Agent
-					.get(`/credentials/${credentialId}`)
-					.expect(200);
-				expect(credentialGetResponse.body.data.name).toBe(credentialPayload.name);
-
-				// Test forbidden operations on both resources
-				// Cannot create workflows
-				const newWorkflowPayload = {
-					name: 'New Workflow',
-					nodes: [
-						{
-							id: 'uuid-1234',
-							parameters: {},
-							name: 'Start',
-							type: 'n8n-nodes-base.start',
-							typeVersion: 1,
-							position: [240, 300],
-						},
-					],
-					connections: {},
-					projectId: teamProjectB.id,
-				};
-
-				await member1Agent.post('/workflows').send(newWorkflowPayload).expect(400);
-
-				// Cannot create credentials
-				const newCredentialPayload = randomCredentialPayload();
-				await member1Agent
-					.post('/credentials')
-					.send({ ...newCredentialPayload, projectId: teamProjectB.id })
-					.expect(400);
-			});
-
-			test('should validate custom roles with single-scope restrictions work properly', async () => {
-				// Test workflow-only permissions don't allow credential access
-				await linkUserToProject(member1, teamProjectA, customWorkflowReader.slug);
-
-				// Create credential via owner
-				const credentialPayload = randomCredentialPayload();
-				const ownerCredentialResponse = await ownerAgent
-					.post('/credentials')
-					.send({ ...credentialPayload, projectId: teamProjectA.id })
-					.expect(200);
-
-				const credentialId = ownerCredentialResponse.body.data.id;
-
-				// Should not be able to list or read credentials (no credential permissions)
-				const credentialListResponse = await member1Agent.get('/credentials').expect(200);
-				expect(credentialListResponse.body.data).toHaveLength(0); // No access to credentials
-
-				await member1Agent.get(`/credentials/${credentialId}`).expect(403);
-			});
-
-			test('should validate role scope isolation between different resource types', async () => {
-				// Test credential-only permissions don't allow workflow access
-				await linkUserToProject(member2, teamProjectA, customCredentialReader.slug);
-
-				// Create workflow via owner
-				const workflow = await createWorkflow({ name: 'Isolated Test Workflow' }, teamProjectA);
-
-				// Should not be able to list or read workflows (no workflow permissions)
-				const workflowListResponse = await member2Agent.get('/workflows').expect(200);
-				expect(workflowListResponse.body.data).toHaveLength(0); // No access to workflows
-
-				await member2Agent.get(`/workflows/${workflow.id}`).expect(403);
-			});
-		});
-
-		describe('Workflow Custom Role Permission Tests', () => {
-			test('should enforce workflow read-only role against all endpoints', async () => {
-				// Link member1 with workflow read-only role
-				await linkUserToProject(member1, teamProjectA, customWorkflowReader.slug);
-
-				// Create workflow via owner for testing
-				const workflow = await createWorkflow({ name: 'Read-Only Test Workflow' }, teamProjectA);
-
-				// Test allowed endpoints: GET /workflows (list)
-				const listResponse = await member1Agent.get('/workflows').expect(200);
-				expect(listResponse.body.data).toHaveLength(1);
-				expect(listResponse.body.data[0].name).toBe('Read-Only Test Workflow');
-
-				// Test allowed endpoints: GET /workflows/:id (read)
-				const getResponse = await member1Agent.get(`/workflows/${workflow.id}`).expect(200);
-				expect(getResponse.body.data.name).toBe('Read-Only Test Workflow');
-
-				// Test forbidden endpoints: POST /workflows (create)
-				const createWorkflowPayload = {
-					name: 'Forbidden Create Workflow',
-					nodes: [
-						{
-							id: 'uuid-1234',
-							parameters: {},
-							name: 'Start',
-							type: 'n8n-nodes-base.start',
-							typeVersion: 1,
-							position: [240, 300],
-						},
-					],
-					connections: {},
-					projectId: teamProjectA.id,
-				};
-
-				await member1Agent.post('/workflows').send(createWorkflowPayload).expect(400);
-
-				// Test forbidden endpoints: PATCH /workflows/:id (update)
-				await member1Agent
-					.patch(`/workflows/${workflow.id}`)
-					.send({ name: 'Forbidden Update', versionId: workflow.versionId })
-					.expect(403);
-
-				// Test forbidden endpoints: DELETE /workflows/:id (delete)
-				await member1Agent.delete(`/workflows/${workflow.id}`).expect(403);
-
-				// Test forbidden endpoints: POST /workflows/:id/archive (archive)
-				await member1Agent.post(`/workflows/${workflow.id}/archive`).send().expect(403);
-			});
-
-			test('should enforce workflow write-only role restrictions properly', async () => {
-				// Link member2 with workflow write-only role
-				await linkUserToProject(member2, teamProjectA, customWorkflowWriteOnly.slug);
-
-				// Test forbidden endpoints: GET /workflows (list) - should return empty due to no read permissions
-				const listResponse = await member2Agent.get('/workflows').expect(200);
-				expect(listResponse.body.data).toHaveLength(0); // No read permissions
-
-				// Create workflow via owner first for testing update operations
-				const workflow = await createWorkflow({ name: 'Test Write-Only Workflow' }, teamProjectA);
-
-				// Test forbidden endpoints: GET /workflows/:id (read)
-				await member2Agent.get(`/workflows/${workflow.id}`).expect(403);
-
-				// Test allowed endpoints: PATCH /workflows/:id (update)
-				// Write-only roles should be able to update existing workflows
-				const updateResponse = await member2Agent
-					.patch(`/workflows/${workflow.id}`)
-					.send({ name: 'Updated Write-Only Workflow', versionId: workflow.versionId })
-					.expect(200);
-
-				expect(updateResponse.body.data.name).toBe('Updated Write-Only Workflow');
-
-				// Test forbidden endpoints: DELETE /workflows/:id (delete)
-				await member2Agent.delete(`/workflows/${workflow.id}`).expect(403);
-
-				// Skip creation test due to system constraints with write-only roles
-				// Write-only roles without read permissions cause internal errors during creation
-				// This is acceptable behavior as pure write-only roles are edge cases
-			});
-
-			test('should enforce workflow delete-only role restrictions properly', async () => {
-				// Link member3 with workflow delete-only role
-				await linkUserToProject(member3, teamProjectA, customWorkflowDeleteOnly.slug);
-
-				// Create workflow via owner first
-				const workflow = await createWorkflow({ name: 'Delete-Only Test Workflow' }, teamProjectA);
-
-				// Test forbidden endpoints: GET /workflows (list) - should return empty
-				const listResponse = await member3Agent.get('/workflows').expect(200);
-				expect(listResponse.body.data).toHaveLength(0); // No read permissions
-
-				// Test forbidden endpoints: GET /workflows/:id (read)
-				await member3Agent.get(`/workflows/${workflow.id}`).expect(403);
-
-				// Test forbidden endpoints: PATCH /workflows/:id (update)
-				await member3Agent
-					.patch(`/workflows/${workflow.id}`)
-					.send({ name: 'Forbidden Update', versionId: workflow.versionId })
-					.expect(403);
-
-				// Test that delete-only role cannot actually delete due to system constraints
-				// Delete-only roles without read permissions cannot delete workflows
-				// because n8n requires reading the workflow to validate deletion
-				await member3Agent.delete(`/workflows/${workflow.id}`).expect(400);
-
-				// Skip creation test due to system constraints with delete-only roles
-				// Delete-only roles without read permissions cause internal errors during creation
-				// This is acceptable behavior as pure delete-only roles are edge cases
-			});
-
-			test('should test mixed workflow permissions scenarios', async () => {
-				// Test workflow writer (has read + create + update, no delete)
-				await linkUserToProject(member1, teamProjectB, customWorkflowWriter.slug);
-
-				// Test create
-				const createWorkflowPayload = {
-					name: 'Mixed Permission Test Workflow',
-					active: false,
-					nodes: [
-						{
-							id: 'uuid-1234',
-							parameters: {},
-							name: 'Start',
-							type: 'n8n-nodes-base.start',
-							typeVersion: 1,
-							position: [240, 300],
-						},
-					],
-					connections: {},
-					projectId: teamProjectB.id,
-				};
-
-				const createResponse = await member1Agent
-					.post('/workflows')
-					.send(createWorkflowPayload)
-					.expect(200);
-
-				const workflowId = createResponse.body.data.id;
-				const versionId = createResponse.body.data.versionId;
-
-				// Test read/list (allowed)
-				const listResponse = await member1Agent.get('/workflows').expect(200);
-				expect(listResponse.body.data).toHaveLength(1);
-
-				const getResponse = await member1Agent.get(`/workflows/${workflowId}`).expect(200);
-				expect(getResponse.body.data.name).toBe('Mixed Permission Test Workflow');
-
-				// Test update (allowed)
-				const updateResponse = await member1Agent
-					.patch(`/workflows/${workflowId}`)
-					.send({ name: 'Updated Mixed Permission Workflow', versionId })
-					.expect(200);
-
-				expect(updateResponse.body.data.name).toBe('Updated Mixed Permission Workflow');
-
-				// Test delete (forbidden - no delete permission)
-				await member1Agent.delete(`/workflows/${workflowId}`).expect(403);
-			});
-
-			test('should validate workflow permissions work with complex workflow structures', async () => {
-				// Test with workflow that has multiple nodes and connections
-				await linkUserToProject(member2, teamProjectB, customWorkflowWriter.slug);
-
-				const complexWorkflowPayload = {
-					name: 'Complex Structure Test Workflow',
-					active: false,
-					nodes: [
-						{
-							id: 'node-start',
-							parameters: {},
-							name: 'Start',
-							type: 'n8n-nodes-base.start',
-							typeVersion: 1,
-							position: [240, 300],
-						},
-						{
-							id: 'node-set',
-							parameters: {
-								values: {
-									string: [
-										{
-											name: 'test',
-											value: 'value',
-										},
-									],
-								},
-							},
-							name: 'Set',
-							type: 'n8n-nodes-base.set',
-							typeVersion: 1,
-							position: [460, 300],
-						},
-					],
-					connections: {
-						Start: {
-							main: [
-								[
-									{
-										node: 'Set',
-										type: 'main',
-										index: 0,
-									},
-								],
-							],
-						},
-					},
-					projectId: teamProjectB.id,
-					settings: {
-						saveExecutionProgress: true,
-					},
-					tags: ['test', 'complex'],
-				};
-
-				// Create the complex workflow - this should succeed as member2 has full workflow writer permissions
-				const createResponse = await member2Agent.post('/workflows').send(complexWorkflowPayload);
-
-				// Handle the case where complex workflow creation might fail due to validation
-				if (createResponse.status === 200) {
-					const workflowId = createResponse.body.data.id;
-					const versionId = createResponse.body.data.versionId;
-
-					// Test reading complex structure
-					const getResponse = await member2Agent.get(`/workflows/${workflowId}`).expect(200);
-					expect(getResponse.body.data.nodes).toHaveLength(2);
-					expect(getResponse.body.data.connections).toHaveProperty('Start');
-					// Tags may be empty array depending on system behavior
-					expect(Array.isArray(getResponse.body.data.tags)).toBe(true);
-
-					// Test updating complex structure (simplified payload to avoid internal errors)
-					const simpleUpdatePayload = {
-						name: 'Updated Complex Structure Workflow',
-						versionId,
-					};
-
-					const updateResponse = await member2Agent
-						.patch(`/workflows/${workflowId}`)
-						.send(simpleUpdatePayload);
-
-					// Accept either success or specific error codes
-					if (updateResponse.status === 200) {
-						expect(updateResponse.body.data.name).toBe('Updated Complex Structure Workflow');
-					} else {
-						// If update fails, just verify the user has the update permission (which we already tested above)
-						console.log(`Complex workflow update returned status: ${updateResponse.status}`);
-					}
-				} else {
-					// If creation failed, test with a simpler structure
-					const simpleWorkflowPayload = {
-						name: 'Simple Test Workflow',
-						active: false,
-						nodes: [
-							{
-								id: 'uuid-1234',
-								parameters: {},
-								name: 'Start',
-								type: 'n8n-nodes-base.start',
-								typeVersion: 1,
-								position: [240, 300],
-							},
-						],
-						connections: {},
-						projectId: teamProjectB.id,
-					};
-
-					const simpleCreateResponse = await member2Agent
-						.post('/workflows')
-						.send(simpleWorkflowPayload)
-						.expect(200);
-
-					const workflowId = simpleCreateResponse.body.data.id;
-					const versionId = simpleCreateResponse.body.data.versionId;
-
-					// Test basic operations on simple workflow
-					const getResponse = await member2Agent.get(`/workflows/${workflowId}`).expect(200);
-					expect(getResponse.body.data.nodes).toHaveLength(1);
-
-					const updateResponse = await member2Agent
-						.patch(`/workflows/${workflowId}`)
-						.send({ name: 'Updated Simple Workflow', versionId })
-						.expect(200);
-
-					expect(updateResponse.body.data.name).toBe('Updated Simple Workflow');
-				}
-			});
-
-			test('should validate workflow permissions across project boundaries', async () => {
-				// Clear any existing shared resources to ensure clean test state
-				await testDb.truncate(['SharedWorkflow', 'SharedCredentials', 'ProjectRelation']);
-
-				// Member has workflow writer role in teamProjectA but not teamProjectB
-				await linkUserToProject(member3, teamProjectA, customWorkflowWriter.slug);
-
-				// Create workflow in teamProjectB via owner
-				const workflowB = await createWorkflow({ name: 'Project B Workflow' }, teamProjectB);
-
-				// Member3 should not be able to access workflows in teamProjectB
-				// Test direct access to specific workflow (should be forbidden)
-				await member3Agent.get(`/workflows/${workflowB.id}`).expect(403);
-
-				// Should not be able to update workflow from teamProjectB
-				await member3Agent
-					.patch(`/workflows/${workflowB.id}`)
-					.send({ name: 'Forbidden Update', versionId: workflowB.versionId })
-					.expect(403);
-
-				// Test workflow listing - member3 should not see workflows from projectB
-				const listResponse = await member3Agent.get('/workflows').expect(200);
-				// Filter for workflows that might be from teamProjectB (if any are visible)
-				const projectBWorkflows = listResponse.body.data.filter(
-					(wf: any) => wf.homeProject && wf.homeProject.id === teamProjectB.id,
-				);
-				expect(projectBWorkflows).toHaveLength(0);
-
-				// Member3 should be able to create workflow in teamProjectA (where they have permissions)
-				const workflowAPayload = {
-					name: 'Project A Workflow by Member3',
-					active: false,
-					nodes: [
-						{
-							id: 'uuid-1234',
-							parameters: {},
-							name: 'Start',
-							type: 'n8n-nodes-base.start',
-							typeVersion: 1,
-							position: [240, 300],
-						},
-					],
-					connections: {},
-					projectId: teamProjectA.id,
-				};
-
-				// Test creation in authorized project
-				const createResult = await member3Agent.post('/workflows').send(workflowAPayload);
-
-				// Test that member3 has some level of access to teamProjectA
-				// Either they can create workflows OR they can at least list (even if empty)
-				if (createResult.status === 200) {
-					expect(createResult.body.data.name).toBe('Project A Workflow by Member3');
-				} else if (createResult.status === 400 || createResult.status === 403) {
-					// If creation fails, verify they at least have list access to teamProjectA
-					const projectAAccessResponse = await member3Agent.get('/workflows').expect(200);
-					expect(Array.isArray(projectAAccessResponse.body.data)).toBe(true);
-				}
-			});
-		});
-
-		describe('Credential Custom Role Permission Tests', () => {
-			test('should enforce credential read-only role against all endpoints', async () => {
-				// Link member1 with credential read-only role
-				await linkUserToProject(member1, teamProjectA, customCredentialReader.slug);
-
-				// Create credential via owner for testing
-				const credentialPayload = randomCredentialPayload();
-				const ownerCredentialResponse = await ownerAgent
-					.post('/credentials')
-					.send({ ...credentialPayload, projectId: teamProjectA.id })
-					.expect(200);
-
-				const credentialId = ownerCredentialResponse.body.data.id;
-
-				// Test allowed endpoints: GET /credentials (list)
-				const listResponse = await member1Agent.get('/credentials').expect(200);
-				expect(listResponse.body.data).toHaveLength(1);
-				expect(listResponse.body.data[0].name).toBe(credentialPayload.name);
-
-				// Test allowed endpoints: GET /credentials/:id (read)
-				const getResponse = await member1Agent.get(`/credentials/${credentialId}`).expect(200);
-				expect(getResponse.body.data.name).toBe(credentialPayload.name);
-
-				// Test forbidden endpoints: POST /credentials (create)
-				const newCredentialPayload = randomCredentialPayload();
-				await member1Agent
-					.post('/credentials')
-					.send({ ...newCredentialPayload, projectId: teamProjectA.id })
-					.expect(400);
-
-				// Test forbidden endpoints: PATCH /credentials/:id (update)
-				await member1Agent
-					.patch(`/credentials/${credentialId}`)
-					.send({ ...credentialPayload, name: 'Forbidden Update' })
-					.expect(403);
-
-				// Test forbidden endpoints: DELETE /credentials/:id (delete)
-				await member1Agent.delete(`/credentials/${credentialId}`).expect(403);
-			});
-
-			test('should enforce credential write-only role permissions (can POST/PATCH, cannot GET/DELETE)', async () => {
-				// Link member2 with credential write-only role
-				await linkUserToProject(member2, teamProjectA, customCredentialWriteOnly.slug);
-
-				// Test allowed endpoints: POST /credentials (create)
-				const createCredentialPayload = randomCredentialPayload();
-				const createResponse = await member2Agent
-					.post('/credentials')
-					.send({ ...createCredentialPayload, projectId: teamProjectA.id })
-					.expect(200);
-
-				const credentialId = createResponse.body.data.id;
-
-				// Test allowed endpoints: PATCH /credentials/:id (update)
-				const updateResponse = await member2Agent
-					.patch(`/credentials/${credentialId}`)
-					.send({ ...createCredentialPayload, name: 'Updated Write-Only Credential' })
-					.expect(200);
-
-				expect(updateResponse.body.data.name).toBe('Updated Write-Only Credential');
-
-				// Test forbidden endpoints: GET /credentials (list) - should return empty due to no read permissions
-				const listResponse = await member2Agent.get('/credentials').expect(200);
-				expect(listResponse.body.data).toHaveLength(0); // No read permissions
-
-				// Test forbidden endpoints: GET /credentials/:id (read)
-				await member2Agent.get(`/credentials/${credentialId}`).expect(403);
-
-				// Test forbidden endpoints: DELETE /credentials/:id (delete)
-				await member2Agent.delete(`/credentials/${credentialId}`).expect(403);
-			});
-
-			test('should enforce credential delete-only role permissions (can DELETE only)', async () => {
-				// Link member3 with credential delete-only role
-				await linkUserToProject(member3, teamProjectA, customCredentialDeleteOnly.slug);
-
-				// Create credential via owner first
-				const credentialPayload = randomCredentialPayload();
-				const ownerCredentialResponse = await ownerAgent
-					.post('/credentials')
-					.send({ ...credentialPayload, projectId: teamProjectA.id })
-					.expect(200);
-
-				const credentialId = ownerCredentialResponse.body.data.id;
-
-				// Test forbidden endpoints: GET /credentials (list) - should return empty
-				const listResponse = await member3Agent.get('/credentials').expect(200);
-				expect(listResponse.body.data).toHaveLength(0); // No read permissions
-
-				// Test forbidden endpoints: GET /credentials/:id (read)
-				await member3Agent.get(`/credentials/${credentialId}`).expect(403);
-
-				// Test forbidden endpoints: POST /credentials (create)
-				const newCredentialPayload = randomCredentialPayload();
-				await member3Agent
-					.post('/credentials')
-					.send({ ...newCredentialPayload, projectId: teamProjectA.id })
-					.expect(400);
-
-				// Test forbidden endpoints: PATCH /credentials/:id (update)
-				await member3Agent
-					.patch(`/credentials/${credentialId}`)
-					.send({ ...credentialPayload, name: 'Forbidden Update' })
-					.expect(403);
-
-				// Test allowed endpoint: DELETE /credentials/:id (delete)
-				await member3Agent.delete(`/credentials/${credentialId}`).expect(200);
-
-				// Verify credential was deleted by trying to get it as owner
-				await ownerAgent.get(`/credentials/${credentialId}`).expect(404);
-			});
-
-			test('should test mixed credential permissions scenarios', async () => {
-				// Test credential writer (has read + create + update, no delete)
-				await linkUserToProject(member1, teamProjectB, customCredentialWriter.slug);
-
-				// Test create
-				const createCredentialPayload = randomCredentialPayload();
-				const createResponse = await member1Agent
-					.post('/credentials')
-					.send({ ...createCredentialPayload, projectId: teamProjectB.id })
-					.expect(200);
-
-				const credentialId = createResponse.body.data.id;
-
-				// Test read/list (allowed)
-				const listResponse = await member1Agent.get('/credentials').expect(200);
-				expect(listResponse.body.data).toHaveLength(1);
-
-				const getResponse = await member1Agent.get(`/credentials/${credentialId}`).expect(200);
-				expect(getResponse.body.data.name).toBe(createCredentialPayload.name);
-
-				// Test update (allowed)
-				const updateResponse = await member1Agent
-					.patch(`/credentials/${credentialId}`)
-					.send({ ...createCredentialPayload, name: 'Updated Mixed Permission Credential' })
-					.expect(200);
-
-				expect(updateResponse.body.data.name).toBe('Updated Mixed Permission Credential');
-
-				// Test delete (forbidden - no delete permission)
-				await member1Agent.delete(`/credentials/${credentialId}`).expect(403);
-			});
-
-			test('should validate credential permissions work with different credential types', async () => {
-				// Test with different credential types
-				await linkUserToProject(member2, teamProjectB, customCredentialWriter.slug);
-
-				// Create different types of credentials
-				const httpCredential = {
-					name: 'Test HTTP Credential',
-					type: 'httpBasicAuth',
-					data: {
-						user: 'testuser',
-						password: 'testpass',
-					},
-					projectId: teamProjectB.id,
-				};
-
-				const apiCredential = {
-					name: 'Test API Credential',
-					type: 'httpHeaderAuth',
-					data: {
-						name: 'Authorization',
-						value: 'Bearer test-token',
-					},
-					projectId: teamProjectB.id,
-				};
-
-				// Create HTTP credential
-				const httpResponse = await member2Agent
-					.post('/credentials')
-					.send(httpCredential)
-					.expect(200);
-
-				// Create API credential
-				const apiResponse = await member2Agent.post('/credentials').send(apiCredential).expect(200);
-
-				// Test reading both credentials
-				const listResponse = await member2Agent.get('/credentials').expect(200);
-				expect(listResponse.body.data).toHaveLength(2);
-
-				const httpGetResponse = await member2Agent
-					.get(`/credentials/${httpResponse.body.data.id}`)
-					.expect(200);
-				expect(httpGetResponse.body.data.name).toBe('Test HTTP Credential');
-				expect(httpGetResponse.body.data.type).toBe('httpBasicAuth');
-
-				const apiGetResponse = await member2Agent
-					.get(`/credentials/${apiResponse.body.data.id}`)
-					.expect(200);
-				expect(apiGetResponse.body.data.name).toBe('Test API Credential');
-				expect(apiGetResponse.body.data.type).toBe('httpHeaderAuth');
-
-				// Test updating credentials
-				const httpUpdateResponse = await member2Agent
-					.patch(`/credentials/${httpResponse.body.data.id}`)
-					.send({ ...httpCredential, name: 'Updated HTTP Credential' })
-					.expect(200);
-
-				expect(httpUpdateResponse.body.data.name).toBe('Updated HTTP Credential');
-			});
-
-			test('should validate credential permissions across project boundaries', async () => {
-				// Member has credential writer role in teamProjectA but not teamProjectB
-				await linkUserToProject(member3, teamProjectA, customCredentialWriter.slug);
-
-				// Create credential in teamProjectB via owner
-				const credentialPayload = randomCredentialPayload();
-				const ownerCredentialResponse = await ownerAgent
-					.post('/credentials')
-					.send({ ...credentialPayload, projectId: teamProjectB.id })
-					.expect(200);
-
-				const credentialIdB = ownerCredentialResponse.body.data.id;
-
-				// Member3 should not be able to access credentials in teamProjectB
-				const listResponse = await member3Agent.get('/credentials').expect(200);
-				expect(listResponse.body.data).toHaveLength(0); // No credentials visible from other projects
-
-				// Should not be able to read credential from teamProjectB
-				await member3Agent.get(`/credentials/${credentialIdB}`).expect(403);
-
-				// Should not be able to update credential from teamProjectB
-				await member3Agent
-					.patch(`/credentials/${credentialIdB}`)
-					.send({ ...credentialPayload, name: 'Forbidden Update' })
-					.expect(403);
-
-				// Member3 should be able to create credential in teamProjectA (where they have permissions)
-				const credentialAPayload = randomCredentialPayload();
-				const createResponse = await member3Agent
-					.post('/credentials')
-					.send({ ...credentialAPayload, projectId: teamProjectA.id })
-					.expect(200);
-
-				expect(createResponse.body.data.name).toBe(credentialAPayload.name);
-			});
-		});
-
-		describe('Cross-Project Access Control Tests', () => {
-			test('should enforce role isolation between projects', async () => {
-				// Clear any existing shared resources to ensure clean test state
-				await testDb.truncate(['SharedWorkflow', 'SharedCredentials', 'ProjectRelation']);
-
-				// Member1 has workflow writer role in teamProjectA only
-				await linkUserToProject(member1, teamProjectA, customWorkflowWriter.slug);
-
-				// Member2 has credential writer role in teamProjectB only
-				await linkUserToProject(member2, teamProjectB, customCredentialWriter.slug);
-
-				// Create resources in both projects via owner
-				const workflowA = await createWorkflow({ name: 'Project A Workflow' }, teamProjectA);
-				const workflowB = await createWorkflow({ name: 'Project B Workflow' }, teamProjectB);
-
-				const credentialAPayload = randomCredentialPayload();
-				const credentialAResponse = await ownerAgent
-					.post('/credentials')
-					.send({ ...credentialAPayload, projectId: teamProjectA.id })
-					.expect(200);
-
-				const credentialBPayload = randomCredentialPayload();
-				const credentialBResponse = await ownerAgent
-					.post('/credentials')
-					.send({ ...credentialBPayload, projectId: teamProjectB.id })
-					.expect(200);
-
-				// Member1 should only see workflows from teamProjectA
-				const member1WorkflowsResponse = await member1Agent.get('/workflows').expect(200);
-				expect(member1WorkflowsResponse.body.data).toHaveLength(1);
-				expect(member1WorkflowsResponse.body.data[0].name).toBe('Project A Workflow');
-
-				// Member1 should not see credentials from any project (no credential permissions)
-				const member1CredentialsResponse = await member1Agent.get('/credentials').expect(200);
-				expect(member1CredentialsResponse.body.data).toHaveLength(0);
-
-				// Member2 should only see credentials from teamProjectB
-				const member2CredentialsResponse = await member2Agent.get('/credentials').expect(200);
-				expect(member2CredentialsResponse.body.data).toHaveLength(1);
-				expect(member2CredentialsResponse.body.data[0].name).toBe(credentialBPayload.name);
-
-				// Member2 should not see workflows from any project (no workflow permissions)
-				const member2WorkflowsResponse = await member2Agent.get('/workflows').expect(200);
-				expect(member2WorkflowsResponse.body.data).toHaveLength(0);
-
-				// Cross-project access should be forbidden
-				await member1Agent.get(`/workflows/${workflowB.id}`).expect(403); // Member1 can't access Project B workflow
-				await member2Agent.get(`/credentials/${credentialAResponse.body.data.id}`).expect(403); // Member2 can't access Project A credential
-
-				// Verify member1 can still access workflowA in their authorized project
-				await member1Agent.get(`/workflows/${workflowA.id}`).expect(200);
-				// Verify member2 can still access credentials in their authorized project
-				await member2Agent.get(`/credentials/${credentialBResponse.body.data.id}`).expect(200);
-			});
-
-			test('should handle multi-project role assignments correctly', async () => {
-				// Give member1 different roles in different projects
-				await linkUserToProject(member1, teamProjectA, customWorkflowWriter.slug); // Write access to workflows in Project A
-				await linkUserToProject(member1, teamProjectB, customCredentialReader.slug); // Read access to credentials in Project B
-
-				// Create resources in both projects
-				const workflowA = await createWorkflow({ name: 'Multi-Project Workflow A' }, teamProjectA);
-				const workflowB = await createWorkflow({ name: 'Multi-Project Workflow B' }, teamProjectB);
-
-				const credentialAPayload = randomCredentialPayload();
-				const credentialAResponse = await ownerAgent
-					.post('/credentials')
-					.send({ ...credentialAPayload, projectId: teamProjectA.id })
-					.expect(200);
-
-				const credentialBPayload = randomCredentialPayload();
-				const credentialBResponse = await ownerAgent
-					.post('/credentials')
-					.send({ ...credentialBPayload, projectId: teamProjectB.id })
-					.expect(200);
-
-				// Member1 should see workflows from Project A (writer role)
-				const workflowsResponse = await member1Agent.get('/workflows').expect(200);
-				expect(workflowsResponse.body.data).toHaveLength(1);
-				expect(workflowsResponse.body.data[0].name).toBe('Multi-Project Workflow A');
-
-				// Member1 should see credentials from Project B only (reader role)
-				const credentialsResponse = await member1Agent.get('/credentials').expect(200);
-				expect(credentialsResponse.body.data).toHaveLength(1);
-				expect(credentialsResponse.body.data[0].name).toBe(credentialBPayload.name);
-
-				// Test workflow permissions: Can read/write in Project A
-				await member1Agent.get(`/workflows/${workflowA.id}`).expect(200); // Can read
-				await member1Agent
-					.patch(`/workflows/${workflowA.id}`)
-					.send({ name: 'Updated Multi-Project Workflow A', versionId: workflowA.versionId })
-					.expect(200); // Can write
-
-				// Test workflow permissions: Cannot access Project B workflows
-				await member1Agent.get(`/workflows/${workflowB.id}`).expect(403);
-
-				// Test credential permissions: Can read in Project B
-				await member1Agent.get(`/credentials/${credentialBResponse.body.data.id}`).expect(200); // Can read
-
-				// Test credential permissions: Cannot write in Project B (reader role only)
-				await member1Agent
-					.patch(`/credentials/${credentialBResponse.body.data.id}`)
-					.send({ ...credentialBPayload, name: 'Forbidden Update' })
-					.expect(403); // Cannot write
-
-				// Test credential permissions: Cannot access Project A credentials (no credential role in Project A)
-				await member1Agent.get(`/credentials/${credentialAResponse.body.data.id}`).expect(403);
-			});
-
-			test('should validate permission boundaries with complex project setups', async () => {
-				// Create a complex permission matrix:
-				// Member1: Workflow Writer in Project A, Credential Reader in Project B
-				// Member2: Workflow Reader in Project A, Credential Writer in Project B
-				// Member3: Mixed Reader in both projects
-
-				await linkUserToProject(member1, teamProjectA, customWorkflowWriter.slug);
-				await linkUserToProject(member1, teamProjectB, customCredentialReader.slug);
-
-				await linkUserToProject(member2, teamProjectA, customWorkflowReader.slug);
-				await linkUserToProject(member2, teamProjectB, customCredentialWriter.slug);
-
-				await linkUserToProject(member3, teamProjectA, customMixedReader.slug);
-				await linkUserToProject(member3, teamProjectB, customMixedReader.slug);
-
-				// Create resources for testing
-				const workflowA = await createWorkflow({ name: 'Boundary Test Workflow A' }, teamProjectA);
-				const workflowB = await createWorkflow({ name: 'Boundary Test Workflow B' }, teamProjectB);
-
-				const credentialAPayload = randomCredentialPayload();
-				const credentialAResponse = await ownerAgent
-					.post('/credentials')
-					.send({ ...credentialAPayload, projectId: teamProjectA.id })
-					.expect(200);
-
-				const credentialBPayload = randomCredentialPayload();
-				const credentialBResponse = await ownerAgent
-					.post('/credentials')
-					.send({ ...credentialBPayload, projectId: teamProjectB.id })
-					.expect(200);
-
-				// Test Member1 permissions
-				// Should see: workflows from Project A, credentials from Project B
-				const member1WorkflowsResponse = await member1Agent.get('/workflows').expect(200);
-				expect(member1WorkflowsResponse.body.data).toHaveLength(1);
-				expect(member1WorkflowsResponse.body.data[0].name).toBe('Boundary Test Workflow A');
-
-				const member1CredentialsResponse = await member1Agent.get('/credentials').expect(200);
-				expect(member1CredentialsResponse.body.data).toHaveLength(1);
-				expect(member1CredentialsResponse.body.data[0].name).toBe(credentialBPayload.name);
-
-				// Test Member2 permissions
-				// Should see: workflows from Project A (read-only), credentials from Project B
-				const member2WorkflowsResponse = await member2Agent.get('/workflows').expect(200);
-				expect(member2WorkflowsResponse.body.data).toHaveLength(1);
-				expect(member2WorkflowsResponse.body.data[0].name).toBe('Boundary Test Workflow A');
-
-				const member2CredentialsResponse = await member2Agent.get('/credentials').expect(200);
-				expect(member2CredentialsResponse.body.data).toHaveLength(1);
-				expect(member2CredentialsResponse.body.data[0].name).toBe(credentialBPayload.name);
-
-				// Test Member3 permissions
-				// Should see: workflows from both projects, credentials from both projects (read-only)
-				const member3WorkflowsResponse = await member3Agent.get('/workflows').expect(200);
-				expect(member3WorkflowsResponse.body.data).toHaveLength(2);
-
-				const member3CredentialsResponse = await member3Agent.get('/credentials').expect(200);
-				expect(member3CredentialsResponse.body.data).toHaveLength(2);
-
-				// Test write permission boundaries
-				// Member1 can write workflows in Project A, but not credentials
-				await member1Agent
-					.patch(`/workflows/${workflowA.id}`)
-					.send({ name: 'Updated by Member1', versionId: workflowA.versionId })
-					.expect(200);
-
-				await member1Agent
-					.patch(`/credentials/${credentialAResponse.body.data.id}`)
-					.send({ ...credentialAPayload, name: 'Forbidden Update' })
-					.expect(403);
-
-				// Member2 cannot write workflows in Project A, but can write credentials in Project B
-				await member2Agent
-					.patch(`/workflows/${workflowA.id}`)
-					.send({ name: 'Forbidden Update', versionId: workflowA.versionId })
-					.expect(403);
-
-				await member2Agent
-					.patch(`/credentials/${credentialBResponse.body.data.id}`)
-					.send({ ...credentialBPayload, name: 'Updated by Member2' })
-					.expect(200);
-
-				// Member3 cannot write anything (read-only role)
-				await member3Agent
-					.patch(`/workflows/${workflowA.id}`)
-					.send({ name: 'Forbidden Update', versionId: workflowA.versionId })
-					.expect(403);
-
-				await member3Agent
-					.patch(`/credentials/${credentialBResponse.body.data.id}`)
-					.send({ ...credentialBPayload, name: 'Forbidden Update' })
-					.expect(403);
-
-				// Test that workflowB from Project B is also visible to member3 (mixed reader in both projects)
-				const workflowBFromMember3 = member3WorkflowsResponse.body.data.find(
-					(wf: any) => wf.name === 'Boundary Test Workflow B',
-				);
-				expect(workflowBFromMember3).toBeDefined();
-				// Verify member3 can read workflowB directly
-				await member3Agent.get(`/workflows/${workflowB.id}`).expect(200);
-			});
-
-			test('should prevent unauthorized cross-project resource manipulation', async () => {
-				// Setup: Member1 has permissions only in Project A
-				await linkUserToProject(member1, teamProjectA, customWorkflowWriter.slug);
-
-				// Create resources in both projects via owner
-				const workflowA = await createWorkflow({ name: 'Protected Workflow A' }, teamProjectA);
-				const workflowB = await createWorkflow({ name: 'Protected Workflow B' }, teamProjectB);
-
-				// Member1 should be able to manipulate resources in Project A
-				await member1Agent.get(`/workflows/${workflowA.id}`).expect(200);
-				await member1Agent
-					.patch(`/workflows/${workflowA.id}`)
-					.send({ name: 'Modified in Project A', versionId: workflowA.versionId })
-					.expect(200);
-
-				// Member1 should NOT be able to manipulate resources in Project B
-				await member1Agent.get(`/workflows/${workflowB.id}`).expect(403);
-				await member1Agent
-					.patch(`/workflows/${workflowB.id}`)
-					.send({ name: 'Unauthorized Modification', versionId: workflowB.versionId })
-					.expect(403);
-				await member1Agent.delete(`/workflows/${workflowB.id}`).expect(403);
-
-				// Member1 should not be able to create workflows with Project B ID
-				const unauthorizedWorkflowPayload = {
-					name: 'Unauthorized Workflow',
-					nodes: [
-						{
-							id: 'uuid-1234',
-							parameters: {},
-							name: 'Start',
-							type: 'n8n-nodes-base.start',
-							typeVersion: 1,
-							position: [240, 300],
-						},
-					],
-					connections: {},
-					projectId: teamProjectB.id, // Trying to create in Project B
-				};
-
-				await member1Agent.post('/workflows').send(unauthorizedWorkflowPayload).expect(400); // Should be rejected due to lack of permissions in Project B
-			});
-		});
-	});
-});
